test: move pure-codec tests from stackslib into stacks-codec
Addresses the PR #7200 reviewer comment about missing tests in
stacks-codec by relocating tests whose dependencies are already
satisfied by stacks-codec + stacks-common + clarity-types.

Moves from stackslib/src/chainstate/stacks/transaction.rs:
- test_transaction_payload_token_transfer (rstest)
- clarity_version_codec_is_consistent (replaces the cross-crate
  test_clarity_versions rstest template with an inline #[case]
  enumeration so stacks-codec does not pick up a clarity dev-dep)
- test_transaction_contract_call_codec
- test_transaction_smart_contract_codec
- test_transaction_payload_versioned_contracts_codec (rstest)
- tx_stacks_transaction_payload_coinbase
- tx_stacks_transaction_payload_nakamoto_coinbase
- tx_stacks_transaction_payload_nakamoto_coinbase_alt_recipient
- tx_stacks_transaction_payload_microblock_poison
- tx_stacks_transaction_payload_invalid (and contract/function name
  variants)
- tx_stacks_asset
- tx_stacks_postcondition (and nft_maybe_sent variant)
- tx_stacks_transaction_codec_originator_mode_and_nft_maybe_sent
- tx_stacks_postcondition_invalid

Moves from stackslib/src/util_lib/strings.rs:
- tx_stacks_strings_codec (renamed stacks_strings_codec)
- tx_stacks_string_invalid (renamed stacks_string_rejects_non_printable)

Supporting changes:
- Add a local copy of check_codec_and_corruption to stacks-codec so
  tests don't reach into stackslib's net codec helpers.
- Add rstest plus testing-feature stacks-common/clarity-types as
  dev-dependencies on stacks-codec.
- Prune now-unused imports in stackslib's test modules.

stackslib retains StacksTransactionSigner, all sign/verify integration
tests, and tx_stacks_transaction_codec (which depends on
codec_all_transactions). 28 tests pass in stacks-codec; the remaining
transaction tests still pass in stackslib.
