+++ stacks-codec/Cargo.toml
@@ -19,6 +19,11 @@ stacks-common = { path = "../stacks-common", default-features = false }
 clarity-types = { path = "../clarity-types", default-features = false }
 serde = { workspace = true, features = ["derive"] }
 
+[dev-dependencies]
+stacks-common = { path = "../stacks-common", default-features = false, features = ["testing"] }
+clarity-types = { path = "../clarity-types", default-features = false, features = ["testing"] }
+rstest = "0.17.0"
+
 [features]
 default = []
 testing = ["stacks-common/testing", "clarity-types/testing"]
