+++ stackslib/src/chainstate/stacks/transaction.rs
@@ -184,15 +184,10 @@ impl StacksTransactionSigner {
 
 #[cfg(test)]
 mod test {
-    use std::io::{Read, Write};
-
     use clarity::types::StacksEpochId;
     use clarity::vm::representations::{ClarityName, ContractName};
-    use clarity::vm::tests::test_clarity_versions;
-    use clarity::vm::types::{PrincipalData, QualifiedContractIdentifier};
-    use clarity::vm::{ClarityVersion, Value};
-    use rstest::rstest;
-    use stacks_common::codec::{read_next, write_next, Error as codec_error, StacksMessageCodec};
+    use clarity::vm::Value;
+    use stacks_common::codec::StacksMessageCodec;
     use stacks_common::types::chainstate::StacksAddress;
     use stacks_common::util::hash::*;
     use stacks_common::util::retry::LogReader;
@@ -200,41 +195,6 @@ mod test {
     use crate::burnchains::Txid;
     use crate::chainstate::stacks::*;
 
-    // Local helpers that mirror the private codec helpers in
-    // `stacks_codec::transaction`; kept here so the tests can construct
-    // expected byte streams directly.
-    #[allow(non_snake_case)]
-    fn ClarityVersion_consensus_serialize<W: Write>(
-        version: &ClarityVersion,
-        fd: &mut W,
-    ) -> Result<(), codec_error> {
-        match *version {
-            ClarityVersion::Clarity1 => write_next(fd, &1u8)?,
-            ClarityVersion::Clarity2 => write_next(fd, &2u8)?,
-            ClarityVersion::Clarity3 => write_next(fd, &3u8)?,
-            ClarityVersion::Clarity4 => write_next(fd, &4u8)?,
-            ClarityVersion::Clarity5 => write_next(fd, &5u8)?,
-        }
-        Ok(())
-    }
-
-    #[allow(non_snake_case)]
-    fn ClarityVersion_consensus_deserialize<R: Read>(
-        fd: &mut R,
-    ) -> Result<ClarityVersion, codec_error> {
-        let version_byte: u8 = read_next(fd)?;
-        match version_byte {
-            1u8 => Ok(ClarityVersion::Clarity1),
-            2u8 => Ok(ClarityVersion::Clarity2),
-            3u8 => Ok(ClarityVersion::Clarity3),
-            4u8 => Ok(ClarityVersion::Clarity4),
-            5u8 => Ok(ClarityVersion::Clarity5),
-            _ => Err(codec_error::DeserializeError(format!(
-                "Unrecognized ClarityVersion byte {}",
-                &version_byte
-            ))),
-        }
-    }
     use crate::chainstate::stacks::test::codec_all_transactions;
     use crate::chainstate::stacks::{
         C32_ADDRESS_VERSION_MAINNET_MULTISIG, C32_ADDRESS_VERSION_MAINNET_SINGLESIG,
@@ -982,1937 +942,6 @@ mod test {
             }
             // restore
             tx_bytes[i] = next_byte as u8;
-        }
-    }
-
-    fn create_token_transfer_bytes(
-        addr: &PrincipalData,
-        amount: u64,
-        memo: &TokenTransferMemo,
-    ) -> Vec<u8> {
-        let mut bytes = vec![];
-        bytes.push(TransactionPayloadID::TokenTransfer as u8);
-        addr.consensus_serialize(&mut bytes).unwrap();
-        bytes.extend_from_slice(&amount.to_be_bytes());
-        bytes.extend_from_slice(&memo.0);
-        bytes
-    }
-
-    #[rstest]
-    #[case::standard_address(
-        PrincipalData::from(StacksAddress::new(1, Hash160([0xff; 20])).unwrap())
-    )]
-    #[case::contract_address(
-        PrincipalData::from(QualifiedContractIdentifier {
-            issuer: StacksAddress::new(1, Hash160([0xff; 20])).unwrap().into(),
-            name: ContractName::from_literal("foo-contract"),
-        })
-    )]
-    fn test_transaction_payload_token_transfer(#[case] addr: PrincipalData) {
-        let memo = TokenTransferMemo([1u8; 34]);
-        let amount = 123u64;
-        let payload = TransactionPayload::TokenTransfer(addr.clone(), amount, memo.clone());
-        let expected_bytes = create_token_transfer_bytes(&addr, amount, &memo);
-        check_codec_and_corruption::<TransactionPayload>(&payload, &expected_bytes);
-    }
-
-    #[apply(test_clarity_versions)]
-    fn clarity_version_codec_is_consistent(
-        #[case] version: ClarityVersion,
-        #[case] _epoch: StacksEpochId,
-    ) {
-        use std::io::Cursor;
-
-        let mut buf = vec![];
-        ClarityVersion_consensus_serialize(&version, &mut buf).unwrap();
-        let mut cursor = Cursor::new(&buf);
-        let decoded = ClarityVersion_consensus_deserialize(&mut cursor).unwrap();
-        assert_eq!(version, decoded, "Roundtrip mismatch for {:?}", version);
-    }
-
-    // Helper functions for test data creation
-    fn sample_contract_call() -> TransactionContractCall {
-        TransactionContractCall {
-            address: StacksAddress::new(1, Hash160([0xff; 20])).unwrap(),
-            contract_name: ContractName::try_from("hello-contract-name").unwrap(),
-            function_name: ClarityName::try_from("hello-function-name").unwrap(),
-            function_args: vec![Value::Int(0)],
-        }
-    }
-
-    fn sample_smart_contract() -> TransactionSmartContract {
-        TransactionSmartContract {
-            name: ContractName::try_from("hello-contract-name").unwrap(),
-            code_body: StacksString::from_str("hello contract code body").unwrap(),
-        }
-    }
-
-    fn serialize_contract_call(contract_call: &TransactionContractCall) -> Vec<u8> {
-        let mut bytes = vec![];
-        contract_call
-            .address
-            .consensus_serialize(&mut bytes)
-            .unwrap();
-        contract_call
-            .contract_name
-            .consensus_serialize(&mut bytes)
-            .unwrap();
-        contract_call
-            .function_name
-            .consensus_serialize(&mut bytes)
-            .unwrap();
-        contract_call
-            .function_args
-            .consensus_serialize(&mut bytes)
-            .unwrap();
-        bytes
-    }
-
-    fn serialize_smart_contract(smart_contract: &TransactionSmartContract) -> Vec<u8> {
-        let mut bytes = vec![];
-        smart_contract.name.consensus_serialize(&mut bytes).unwrap();
-        smart_contract
-            .code_body
-            .consensus_serialize(&mut bytes)
-            .unwrap();
-        bytes
-    }
-
-    fn serialize_versioned_smart_contract(
-        smart_contract: &TransactionSmartContract,
-        version: &ClarityVersion,
-    ) -> Vec<u8> {
-        let mut bytes = vec![];
-        ClarityVersion_consensus_serialize(version, &mut bytes).unwrap();
-        smart_contract.name.consensus_serialize(&mut bytes).unwrap();
-        smart_contract
-            .code_body
-            .consensus_serialize(&mut bytes)
-            .unwrap();
-        bytes
-    }
-
-    fn create_transaction_payload_bytes(
-        payload_id: TransactionPayloadID,
-        mut payload_bytes: Vec<u8>,
-    ) -> Vec<u8> {
-        let mut bytes = vec![payload_id as u8];
-        bytes.append(&mut payload_bytes);
-        bytes
-    }
-
-    #[test]
-    fn test_transaction_contract_call_codec() {
-        let contract_call = sample_contract_call();
-        let expected_bytes = serialize_contract_call(&contract_call);
-        check_codec_and_corruption::<TransactionContractCall>(&contract_call, &expected_bytes);
-    }
-
-    #[test]
-    fn test_transaction_smart_contract_codec() {
-        let smart_contract = sample_smart_contract();
-        let expected_bytes = serialize_smart_contract(&smart_contract);
-        check_codec_and_corruption::<TransactionSmartContract>(&smart_contract, &expected_bytes);
-    }
-
-    #[apply(test_clarity_versions)]
-    fn test_transaction_payload_versioned_contracts_codec(
-        #[case] version: ClarityVersion,
-        _epoch: StacksEpochId,
-    ) {
-        let payload = TransactionPayload::SmartContract(sample_smart_contract(), Some(version));
-        let expected_bytes = create_transaction_payload_bytes(
-            TransactionPayloadID::VersionedSmartContract,
-            serialize_versioned_smart_contract(&sample_smart_contract(), &version),
-        );
-        check_codec_and_corruption::<TransactionPayload>(&payload, &expected_bytes);
-    }
-
-    #[test]
-    fn tx_stacks_transaction_payload_coinbase() {
-        let coinbase_payload =
-            TransactionPayload::Coinbase(CoinbasePayload([0x12; 32]), None, None);
-        let coinbase_payload_bytes = vec![
-            // payload type ID
-            TransactionPayloadID::Coinbase as u8,
-            // buffer
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-        ];
-
-        check_codec_and_corruption::<TransactionPayload>(
-            &coinbase_payload,
-            &coinbase_payload_bytes,
-        );
-    }
-
-    #[test]
-    fn tx_stacks_transaction_payload_nakamoto_coinbase() {
-        let proof_bytes = hex_bytes("9275df67a68c8745c0ff97b48201ee6db447f7c93b23ae24cdc2400f52fdb08a1a6ac7ec71bf9c9c76e96ee4675ebff60625af28718501047bfd87b810c2d2139b73c23bd69de66360953a642c2a330a").unwrap();
-        let proof = VRFProof::from_bytes(&proof_bytes[..]).unwrap();
-
-        let coinbase_payload =
-            TransactionPayload::Coinbase(CoinbasePayload([0x12; 32]), None, Some(proof));
-        let coinbase_bytes = vec![
-            // payload type ID
-            TransactionPayloadID::NakamotoCoinbase as u8,
-            // buffer
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            // no alt recipient, so Value::none
-            0x09,
-            // proof bytes
-            0x92,
-            0x75,
-            0xdf,
-            0x67,
-            0xa6,
-            0x8c,
-            0x87,
-            0x45,
-            0xc0,
-            0xff,
-            0x97,
-            0xb4,
-            0x82,
-            0x01,
-            0xee,
-            0x6d,
-            0xb4,
-            0x47,
-            0xf7,
-            0xc9,
-            0x3b,
-            0x23,
-            0xae,
-            0x24,
-            0xcd,
-            0xc2,
-            0x40,
-            0x0f,
-            0x52,
-            0xfd,
-            0xb0,
-            0x8a,
-            0x1a,
-            0x6a,
-            0xc7,
-            0xec,
-            0x71,
-            0xbf,
-            0x9c,
-            0x9c,
-            0x76,
-            0xe9,
-            0x6e,
-            0xe4,
-            0x67,
-            0x5e,
-            0xbf,
-            0xf6,
-            0x06,
-            0x25,
-            0xaf,
-            0x28,
-            0x71,
-            0x85,
-            0x01,
-            0x04,
-            0x7b,
-            0xfd,
-            0x87,
-            0xb8,
-            0x10,
-            0xc2,
-            0xd2,
-            0x13,
-            0x9b,
-            0x73,
-            0xc2,
-            0x3b,
-            0xd6,
-            0x9d,
-            0xe6,
-            0x63,
-            0x60,
-            0x95,
-            0x3a,
-            0x64,
-            0x2c,
-            0x2a,
-            0x33,
-            0x0a,
-        ];
-
-        check_codec_and_corruption(&coinbase_payload, &coinbase_bytes);
-    }
-
-    #[test]
-    fn tx_stacks_transaction_payload_nakamoto_coinbase_alt_recipient() {
-        let proof_bytes = hex_bytes("9275df67a68c8745c0ff97b48201ee6db447f7c93b23ae24cdc2400f52fdb08a1a6ac7ec71bf9c9c76e96ee4675ebff60625af28718501047bfd87b810c2d2139b73c23bd69de66360953a642c2a330a").unwrap();
-        let proof = VRFProof::from_bytes(&proof_bytes[..]).unwrap();
-
-        let recipient = PrincipalData::from(QualifiedContractIdentifier {
-            issuer: StacksAddress::new(1, Hash160([0xff; 20])).unwrap().into(),
-            name: ContractName::from_literal("foo-contract"),
-        });
-
-        let coinbase_payload =
-            TransactionPayload::Coinbase(CoinbasePayload([0x12; 32]), Some(recipient), Some(proof));
-        let coinbase_bytes = vec![
-            // payload type ID
-            TransactionPayloadID::NakamotoCoinbase as u8,
-            // buffer
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            0x12,
-            // have contract recipient, so Some(..)
-            0x0a,
-            // contract address type
-            0x06,
-            // address
-            0x01,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            0xff,
-            // name length
-            0x0c,
-            // name ('foo-contract')
-            0x66,
-            0x6f,
-            0x6f,
-            0x2d,
-            0x63,
-            0x6f,
-            0x6e,
-            0x74,
-            0x72,
-            0x61,
-            0x63,
-            0x74,
-            // proof bytes
-            0x92,
-            0x75,
-            0xdf,
-            0x67,
-            0xa6,
-            0x8c,
-            0x87,
-            0x45,
-            0xc0,
-            0xff,
-            0x97,
-            0xb4,
-            0x82,
-            0x01,
-            0xee,
-            0x6d,
-            0xb4,
-            0x47,
-            0xf7,
-            0xc9,
-            0x3b,
-            0x23,
-            0xae,
-            0x24,
-            0xcd,
-            0xc2,
-            0x40,
-            0x0f,
-            0x52,
-            0xfd,
-            0xb0,
-            0x8a,
-            0x1a,
-            0x6a,
-            0xc7,
-            0xec,
-            0x71,
-            0xbf,
-            0x9c,
-            0x9c,
-            0x76,
-            0xe9,
-            0x6e,
-            0xe4,
-            0x67,
-            0x5e,
-            0xbf,
-            0xf6,
-            0x06,
-            0x25,
-            0xaf,
-            0x28,
-            0x71,
-            0x85,
-            0x01,
-            0x04,
-            0x7b,
-            0xfd,
-            0x87,
-            0xb8,
-            0x10,
-            0xc2,
-            0xd2,
-            0x13,
-            0x9b,
-            0x73,
-            0xc2,
-            0x3b,
-            0xd6,
-            0x9d,
-            0xe6,
-            0x63,
-            0x60,
-            0x95,
-            0x3a,
-            0x64,
-            0x2c,
-            0x2a,
-            0x33,
-            0x0a,
-        ];
-
-        check_codec_and_corruption(&coinbase_payload, &coinbase_bytes);
-    }
-
-    #[test]
-    fn tx_stacks_transaction_payload_microblock_poison() {
-        let header_1 = StacksMicroblockHeader {
-            version: 0x12,
-            sequence: 0x34,
-            prev_block: EMPTY_MICROBLOCK_PARENT_HASH.clone(),
-            tx_merkle_root: Sha512Trunc256Sum([1u8; 32]),
-            signature: MessageSignature([2u8; 65]),
-        };
-
-        let header_2 = StacksMicroblockHeader {
-            version: 0x12,
-            sequence: 0x34,
-            prev_block: EMPTY_MICROBLOCK_PARENT_HASH.clone(),
-            tx_merkle_root: Sha512Trunc256Sum([2u8; 32]),
-            signature: MessageSignature([3u8; 65]),
-        };
-
-        let payload = TransactionPayload::PoisonMicroblock(header_1, header_2);
-
-        let payload_bytes = vec![
-            // payload type ID
-            TransactionPayloadID::PoisonMicroblock as u8,
-            // header_1
-            // version
-            0x12,
-            // sequence
-            0x00,
-            0x34,
-            // prev block
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            // tx merkle root
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            // signature
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            // header_2
-            // version
-            0x12,
-            // sequence
-            0x00,
-            0x34,
-            // prev block
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            // tx merkle root
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            // signature
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-        ];
-
-        check_codec_and_corruption::<TransactionPayload>(&payload, &payload_bytes);
-
-        let payload_bytes_bad_parent = vec![
-            // payload type ID
-            TransactionPayloadID::PoisonMicroblock as u8,
-            // header_1
-            // version
-            0x12,
-            // sequence
-            0x00,
-            0x35,
-            // prev block
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            // tx merkle root
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            0x01,
-            // signature
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            // header_2
-            // version
-            0x12,
-            // sequence
-            0x00,
-            0x34,
-            // prev block
-            0x01,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            // tx merkle root
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            // signature
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-            0x03,
-        ];
-
-        assert!(
-            TransactionPayload::consensus_deserialize(&mut &payload_bytes_bad_parent[..])
-                .unwrap_err()
-                .to_string()
-                .find("microblock headers do not identify a fork")
-                .is_some()
-        );
-
-        let payload_bytes_equal = vec![
-            // payload type ID
-            TransactionPayloadID::PoisonMicroblock as u8,
-            // header_1
-            // version
-            0x12,
-            // sequence
-            0x00,
-            0x34,
-            // prev block
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            // tx merkle root
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            // signature
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            // header_2
-            // version
-            0x12,
-            // sequence
-            0x00,
-            0x34,
-            // prev block
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            // tx merkle root
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            // signature
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-            0x02,
-        ];
-
-        assert!(
-            TransactionPayload::consensus_deserialize(&mut &payload_bytes_equal[..])
-                .unwrap_err()
-                .to_string()
-                .find("microblock headers match")
-                .is_some()
-        );
-    }
-
-    #[test]
-    fn tx_stacks_transaction_payload_invalid() {
-        let hello_contract_call = "hello-contract-call";
-        let hello_contract_name = "hello-contract-name";
-        let hello_function_name = "hello-function-name";
-
-        let contract_call = TransactionContractCall {
-            address: StacksAddress::new(1, Hash160([0xff; 20])).unwrap(),
-            contract_name: ContractName::try_from(hello_contract_name).unwrap(),
-            function_name: ClarityName::try_from(hello_function_name).unwrap(),
-            function_args: vec![Value::Int(0)],
-        };
-
-        let mut contract_call_bytes = vec![];
-        contract_call
-            .address
-            .consensus_serialize(&mut contract_call_bytes)
-            .unwrap();
-        contract_call
-            .contract_name
-            .consensus_serialize(&mut contract_call_bytes)
-            .unwrap();
-        contract_call
-            .function_name
-            .consensus_serialize(&mut contract_call_bytes)
-            .unwrap();
-        contract_call
-            .function_args
-            .consensus_serialize(&mut contract_call_bytes)
-            .unwrap();
-
-        let mut transaction_contract_call = vec![0xff];
-        transaction_contract_call.append(&mut contract_call_bytes.clone());
-
-        assert!(
-            TransactionPayload::consensus_deserialize(&mut &transaction_contract_call[..])
-                .unwrap_err()
-                .to_string()
-                .find("unknown payload ID")
-                .is_some()
-        );
-    }
-
-    #[test]
-    fn tx_stacks_transaction_payload_invalid_contract_name() {
-        // test invalid contract name
-        let address = StacksAddress::new(1, Hash160([0xff; 20])).unwrap();
-        let contract_name = "hello\x00contract-name";
-        let function_name = ClarityName::try_from("hello-function-name").unwrap();
-        let function_args = vec![Value::Int(0)];
-
-        let mut contract_call_bytes = vec![];
-        address
-            .consensus_serialize(&mut contract_call_bytes)
-            .unwrap();
-        contract_call_bytes.push(contract_name.len() as u8);
-        contract_call_bytes.extend_from_slice(contract_name.as_bytes());
-        function_name
-            .consensus_serialize(&mut contract_call_bytes)
-            .unwrap();
-        function_args
-            .consensus_serialize(&mut contract_call_bytes)
-            .unwrap();
-
-        let mut transaction_contract_call = vec![TransactionPayloadID::ContractCall as u8];
-        transaction_contract_call.append(&mut contract_call_bytes);
-
-        assert!(
-            TransactionPayload::consensus_deserialize(&mut &transaction_contract_call[..])
-                .unwrap_err()
-                .to_string()
-                .find("Failed to parse Contract name")
-                .is_some()
-        );
-    }
-
-    #[test]
-    fn tx_stacks_transaction_payload_invalid_function_name() {
-        // test invalid contract name
-        let address = StacksAddress::new(1, Hash160([0xff; 20])).unwrap();
-        let contract_name = ContractName::try_from("hello-contract-name").unwrap();
-        let hello_function_name = "hello\x00function-name";
-        let mut hello_function_name_bytes = vec![hello_function_name.len() as u8];
-        hello_function_name_bytes.extend_from_slice(hello_function_name.as_bytes());
-
-        let function_args = vec![Value::Int(0)];
-
-        let mut contract_call_bytes = vec![];
-        address
-            .consensus_serialize(&mut contract_call_bytes)
-            .unwrap();
-        contract_name
-            .consensus_serialize(&mut contract_call_bytes)
-            .unwrap();
-        contract_call_bytes.extend_from_slice(&hello_function_name_bytes);
-        function_args
-            .consensus_serialize(&mut contract_call_bytes)
-            .unwrap();
-
-        let mut transaction_contract_call = vec![TransactionPayloadID::ContractCall as u8];
-        transaction_contract_call.append(&mut contract_call_bytes);
-
-        assert!(
-            TransactionPayload::consensus_deserialize(&mut &transaction_contract_call[..])
-                .unwrap_err()
-                .to_string()
-                .find("Failed to parse Clarity name")
-                .is_some()
-        );
-    }
-
-    #[test]
-    fn tx_stacks_asset() {
-        let addr = StacksAddress::new(1, Hash160([0xff; 20])).unwrap();
-        let addr_bytes = [
-            // version
-            0x01, // bytes
-            0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-            0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-        ];
-
-        let asset_name = ClarityName::try_from("hello-asset").unwrap();
-        let mut asset_name_bytes = vec![
-            // length
-            asset_name.len(),
-        ];
-        asset_name_bytes.extend_from_slice(asset_name.to_string().as_str().as_bytes());
-
-        let contract_name = ContractName::try_from("hello-world").unwrap();
-        let mut contract_name_bytes = vec![
-            // length
-            contract_name.len(),
-        ];
-        contract_name_bytes.extend_from_slice(contract_name.to_string().as_str().as_bytes());
-
-        let asset_info = AssetInfo {
-            contract_address: addr.clone(),
-            contract_name,
-            asset_name,
-        };
-
-        let mut asset_info_bytes = vec![];
-        asset_info_bytes.extend_from_slice(&addr_bytes[..]);
-        asset_info_bytes.extend_from_slice(&contract_name_bytes[..]);
-        asset_info_bytes.extend_from_slice(&asset_name_bytes[..]);
-
-        let mut actual_asset_info_bytes = vec![];
-        asset_info
-            .consensus_serialize(&mut actual_asset_info_bytes)
-            .unwrap();
-        assert_eq!(actual_asset_info_bytes, asset_info_bytes);
-
-        assert_eq!(
-            AssetInfo::consensus_deserialize(&mut &asset_info_bytes[..]).unwrap(),
-            asset_info
-        );
-    }
-
-    #[test]
-    fn tx_stacks_postcondition() {
-        let tx_post_condition_principals = vec![
-            PostConditionPrincipal::Origin,
-            PostConditionPrincipal::Standard(StacksAddress::new(1, Hash160([1u8; 20])).unwrap()),
-            PostConditionPrincipal::Contract(
-                StacksAddress::new(2, Hash160([2u8; 20])).unwrap(),
-                ContractName::try_from("hello-world").unwrap(),
-            ),
-        ];
-
-        for tx_pcp in tx_post_condition_principals {
-            let addr = StacksAddress::new(1, Hash160([0xff; 20])).unwrap();
-            let asset_name = ClarityName::try_from("hello-asset").unwrap();
-            let contract_name = ContractName::try_from("contract-name").unwrap();
-
-            let stx_pc =
-                TransactionPostCondition::STX(tx_pcp.clone(), FungibleConditionCode::SentGt, 12345);
-            let fungible_pc = TransactionPostCondition::Fungible(
-                tx_pcp.clone(),
-                AssetInfo {
-                    contract_address: addr.clone(),
-                    contract_name: contract_name.clone(),
-                    asset_name: asset_name.clone(),
-                },
-                FungibleConditionCode::SentGt,
-                23456,
-            );
-
-            let nonfungible_pc = TransactionPostCondition::Nonfungible(
-                tx_pcp.clone(),
-                AssetInfo {
-                    contract_address: addr.clone(),
-                    contract_name: contract_name.clone(),
-                    asset_name: asset_name.clone(),
-                },
-                Value::buff_from(vec![0, 1, 2, 3]).unwrap(),
-                NonfungibleConditionCode::NotSent,
-            );
-
-            let mut stx_pc_bytes = vec![];
-            (AssetInfoID::STX as u8)
-                .consensus_serialize(&mut stx_pc_bytes)
-                .unwrap();
-            tx_pcp.consensus_serialize(&mut stx_pc_bytes).unwrap();
-            stx_pc_bytes.append(&mut vec![
-                // condition code
-                FungibleConditionCode::SentGt as u8,
-                // amount
-                0x00,
-                0x00,
-                0x00,
-                0x00,
-                0x00,
-                0x00,
-                0x30,
-                0x39,
-            ]);
-
-            let mut fungible_pc_bytes = vec![];
-            (AssetInfoID::FungibleAsset as u8)
-                .consensus_serialize(&mut fungible_pc_bytes)
-                .unwrap();
-            tx_pcp.consensus_serialize(&mut fungible_pc_bytes).unwrap();
-            AssetInfo {
-                contract_address: addr.clone(),
-                contract_name: contract_name.clone(),
-                asset_name: asset_name.clone(),
-            }
-            .consensus_serialize(&mut fungible_pc_bytes)
-            .unwrap();
-            fungible_pc_bytes.append(&mut vec![
-                // condition code
-                FungibleConditionCode::SentGt as u8,
-                // amount
-                0x00,
-                0x00,
-                0x00,
-                0x00,
-                0x00,
-                0x00,
-                0x5b,
-                0xa0,
-            ]);
-
-            let mut nonfungible_pc_bytes = vec![];
-            (AssetInfoID::NonfungibleAsset as u8)
-                .consensus_serialize(&mut nonfungible_pc_bytes)
-                .unwrap();
-            tx_pcp
-                .consensus_serialize(&mut nonfungible_pc_bytes)
-                .unwrap();
-            AssetInfo {
-                contract_address: addr.clone(),
-                contract_name: contract_name.clone(),
-                asset_name: asset_name.clone(),
-            }
-            .consensus_serialize(&mut nonfungible_pc_bytes)
-            .unwrap();
-            Value::buff_from(vec![0, 1, 2, 3])
-                .unwrap()
-                .consensus_serialize(&mut nonfungible_pc_bytes)
-                .unwrap();
-            nonfungible_pc_bytes.append(&mut vec![
-                // condition code
-                NonfungibleConditionCode::NotSent as u8,
-            ]);
-
-            let pcs = vec![stx_pc, fungible_pc, nonfungible_pc];
-            let pc_bytes = [stx_pc_bytes, fungible_pc_bytes, nonfungible_pc_bytes];
-            for i in 0..3 {
-                check_codec_and_corruption::<TransactionPostCondition>(&pcs[i], &pc_bytes[i]);
-            }
-        }
-    }
-
-    #[test]
-    fn tx_stacks_postcondition_nft_maybe_sent_codec() {
-        let postcondition = TransactionPostCondition::Nonfungible(
-            PostConditionPrincipal::Origin,
-            AssetInfo {
-                contract_address: StacksAddress::new(1, Hash160([0x11; 20])).unwrap(),
-                contract_name: ContractName::try_from("contract-name").unwrap(),
-                asset_name: ClarityName::try_from("hello-asset").unwrap(),
-            },
-            Value::buff_from(vec![0, 1, 2, 3]).unwrap(),
-            NonfungibleConditionCode::MaybeSent,
-        );
-
-        let mut postcondition_bytes = vec![];
-        postcondition
-            .consensus_serialize(&mut postcondition_bytes)
-            .unwrap();
-
-        #[rustfmt::skip]
-        let expected_bytes = vec![
-            // asset info id
-            0x02,
-            // principal id (origin)
-            0x01,
-            // contract address (version 1, Hash160([0x11; 20]))
-            0x01,
-            0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11,
-            0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11,
-            // contract name "contract-name"
-            0x0d, b'c', b'o', b'n', b't', b'r', b'a', b'c', b't', b'-', b'n', b'a', b'm', b'e',
-            // asset name "hello-asset"
-            0x0b, b'h', b'e', b'l', b'l', b'o', b'-', b'a', b's', b's', b'e', b't',
-            // clarity value: buffer (type prefix 0x02, length 4, data [0,1,2,3])
-            0x02, 0x00, 0x00, 0x00, 0x04, 0x00, 0x01, 0x02, 0x03,
-            // condition code (MaybeSent)
-            0x12,
-        ];
-        assert_eq!(postcondition_bytes, expected_bytes);
-
-        check_codec_and_corruption::<TransactionPostCondition>(
-            &postcondition,
-            &postcondition_bytes,
-        );
-    }
-
-    #[test]
-    fn tx_stacks_transaction_codec_originator_mode_and_nft_maybe_sent() {
-        let auth = TransactionAuth::from_p2pkh(&StacksPrivateKey::random()).unwrap();
-        let mut tx = StacksTransaction::new(
-            TransactionVersion::Testnet,
-            auth,
-            TransactionPayload::new_contract_call(
-                StacksAddress::new(1, Hash160([0x22; 20])).unwrap(),
-                "hello",
-                "world",
-                vec![Value::Int(1)],
-            )
-            .unwrap(),
-        );
-
-        tx.post_condition_mode = TransactionPostConditionMode::Originator;
-        tx.post_conditions
-            .push(TransactionPostCondition::Nonfungible(
-                PostConditionPrincipal::Origin,
-                AssetInfo {
-                    contract_address: StacksAddress::new(1, Hash160([0x33; 20])).unwrap(),
-                    contract_name: ContractName::try_from("contract-name").unwrap(),
-                    asset_name: ClarityName::try_from("hello-asset").unwrap(),
-                },
-                Value::buff_from(vec![4, 5, 6, 7]).unwrap(),
-                NonfungibleConditionCode::MaybeSent,
-            ));
-
-        let mut tx_bytes = vec![];
-        tx.consensus_serialize(&mut tx_bytes).unwrap();
-
-        // Check the post-condition bytes directly within the serialized transaction
-        #[rustfmt::skip]
-        let expected_pc_bytes: &[u8] = &[
-            // post-condition mode (Originator)
-            0x03,
-            // post-conditions length prefix (1 item)
-            0x00, 0x00, 0x00, 0x01,
-            // asset info id (NonfungibleAsset)
-            0x02,
-            // principal id (origin)
-            0x01,
-            // contract address (version 1, Hash160([0x33; 20]))
-            0x01,
-            0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33,
-            0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33,
-            // contract name "contract-name"
-            0x0d, b'c', b'o', b'n', b't', b'r', b'a', b'c', b't', b'-', b'n', b'a', b'm', b'e',
-            // asset name "hello-asset"
-            0x0b, b'h', b'e', b'l', b'l', b'o', b'-', b'a', b's', b's', b'e', b't',
-            // clarity value: buffer (type prefix 0x02, length 4, data [4,5,6,7])
-            0x02, 0x00, 0x00, 0x00, 0x04, 0x04, 0x05, 0x06, 0x07,
-            // condition code (MaybeSent)
-            0x12,
-        ];
-        assert!(
-            tx_bytes
-                .windows(expected_pc_bytes.len())
-                .any(|w| w == expected_pc_bytes),
-            "Expected post-condition bytes not found in serialized transaction"
-        );
-
-        check_codec_and_corruption::<StacksTransaction>(&tx, &tx_bytes);
-    }
-
-    #[test]
-    fn tx_stacks_postcondition_invalid() {
-        let addr = StacksAddress::new(1, Hash160([0xff; 20])).unwrap();
-        let asset_name = ClarityName::try_from("hello-asset").unwrap();
-        let contract_name = ContractName::try_from("hello-world").unwrap();
-
-        // can't parse a postcondition with an invalid condition code
-
-        let mut stx_pc_bytes_bad_condition = vec![];
-        (AssetInfoID::STX as u8)
-            .consensus_serialize(&mut stx_pc_bytes_bad_condition)
-            .unwrap();
-        stx_pc_bytes_bad_condition.append(&mut vec![
-            // principal
-            PostConditionPrincipalID::Origin as u8,
-            // condition code
-            NonfungibleConditionCode::NotSent as u8,
-            // amount
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x30,
-            0x39,
-        ]);
-
-        let mut fungible_pc_bytes_bad_condition = vec![];
-        (AssetInfoID::FungibleAsset as u8)
-            .consensus_serialize(&mut fungible_pc_bytes_bad_condition)
-            .unwrap();
-        fungible_pc_bytes_bad_condition.append(&mut vec![PostConditionPrincipalID::Origin as u8]);
-        AssetInfo {
-            contract_address: addr.clone(),
-            contract_name: contract_name.clone(),
-            asset_name: asset_name.clone(),
-        }
-        .consensus_serialize(&mut fungible_pc_bytes_bad_condition)
-        .unwrap();
-        fungible_pc_bytes_bad_condition.append(&mut vec![
-            // condition code
-            NonfungibleConditionCode::Sent as u8,
-            // amount
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x5b,
-            0xa0,
-        ]);
-
-        let mut nonfungible_pc_bytes_bad_condition = vec![];
-        (AssetInfoID::NonfungibleAsset as u8)
-            .consensus_serialize(&mut nonfungible_pc_bytes_bad_condition)
-            .unwrap();
-        nonfungible_pc_bytes_bad_condition
-            .append(&mut vec![PostConditionPrincipalID::Origin as u8]);
-        AssetInfo {
-            contract_address: addr.clone(),
-            contract_name: contract_name.clone(),
-            asset_name: asset_name.clone(),
-        }
-        .consensus_serialize(&mut nonfungible_pc_bytes_bad_condition)
-        .unwrap();
-        Value::buff_from(vec![0, 1, 2, 3])
-            .unwrap()
-            .consensus_serialize(&mut nonfungible_pc_bytes_bad_condition)
-            .unwrap();
-        nonfungible_pc_bytes_bad_condition.append(&mut vec![
-            // condition code
-            FungibleConditionCode::SentGt as u8,
-        ]);
-
-        let bad_pc_bytes = [
-            stx_pc_bytes_bad_condition,
-            fungible_pc_bytes_bad_condition,
-            nonfungible_pc_bytes_bad_condition,
-        ];
-        for i in 0..3 {
-            assert!(
-                TransactionPostCondition::consensus_deserialize(&mut &bad_pc_bytes[i][..]).is_err()
-            );
-        }
-
-        // can't parse a postcondition with an invalid principal
-
-        let mut stx_pc_bytes_bad_principal = vec![];
-        (AssetInfoID::STX as u8)
-            .consensus_serialize(&mut stx_pc_bytes_bad_principal)
-            .unwrap();
-        stx_pc_bytes_bad_principal.append(&mut vec![
-            // principal
-            0xff,
-            // condition code
-            NonfungibleConditionCode::NotSent as u8,
-            // amount
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x30,
-            0x39,
-        ]);
-
-        let mut fungible_pc_bytes_bad_principal = vec![];
-        (AssetInfoID::FungibleAsset as u8)
-            .consensus_serialize(&mut fungible_pc_bytes_bad_principal)
-            .unwrap();
-        fungible_pc_bytes_bad_principal.append(&mut vec![0xff]);
-        AssetInfo {
-            contract_address: addr.clone(),
-            contract_name: contract_name.clone(),
-            asset_name: asset_name.clone(),
-        }
-        .consensus_serialize(&mut fungible_pc_bytes_bad_principal)
-        .unwrap();
-        fungible_pc_bytes_bad_principal.append(&mut vec![
-            // condition code
-            NonfungibleConditionCode::Sent as u8,
-            // amount
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x00,
-            0x5b,
-            0xa0,
-        ]);
-
-        let mut nonfungible_pc_bytes_bad_principal = vec![];
-        (AssetInfoID::NonfungibleAsset as u8)
-            .consensus_serialize(&mut nonfungible_pc_bytes_bad_principal)
-            .unwrap();
-        nonfungible_pc_bytes_bad_principal.append(&mut vec![0xff]);
-        AssetInfo {
-            contract_address: addr.clone(),
-            contract_name,
-            asset_name,
-        }
-        .consensus_serialize(&mut nonfungible_pc_bytes_bad_principal)
-        .unwrap();
-        Value::buff_from(vec![0, 1, 2, 3])
-            .unwrap()
-            .consensus_serialize(&mut nonfungible_pc_bytes_bad_principal)
-            .unwrap();
-        nonfungible_pc_bytes_bad_principal.append(&mut vec![
-            // condition code
-            FungibleConditionCode::SentGt as u8,
-        ]);
-
-        let bad_pc_bytes = [
-            stx_pc_bytes_bad_principal,
-            fungible_pc_bytes_bad_principal,
-            nonfungible_pc_bytes_bad_principal,
-        ];
-        for i in 0..3 {
-            assert!(
-                TransactionPostCondition::consensus_deserialize(&mut &bad_pc_bytes[i][..]).is_err()
-            );
         }
     }
